# src/admin-ui/contracts/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Dxs.Consigliere",
    "version": "1.0"
  },
  "paths": {
...
    "/api/tx/get/{id}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/tx/state/{id}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionStateResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionStateResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionStateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tx/batch/get": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tx/by-height/get": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "parameters": [
          {
            "name": "blockHeight",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetTransactionsByBlockResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTransactionsByBlockResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTransactionsByBlockResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tx/broadcast": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastTxRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastTxRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastTxRequest"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "429": {
            "description": "Too Many Requests"
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastReceiptDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastReceiptDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastReceiptDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/tx/stas/validate/{id}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "418": {
            "description": "I'm a teapot"
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateStasResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateStasResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateStasResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BroadcastReceiptDto": {
        "type": "object",
        "properties": {
          "txId": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "createdAtMs": {
            "type": "integer",
            "format": "int64"
          },
          "failReason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BroadcastTxRequest": {
        "type": "object",
        "properties": {
          "rawHex": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GetTransactionsByBlockResponse": {
        "type": "object",
        "properties": {
          "blockHeight": {
            "type": "integer",
            "format": "int32"
          },
          "transactions": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "nullable": true
            },
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "TransactionStateResponse": {
        "type": "object",
        "properties": {
          "txId": {
            "type": "string",
            "nullable": true
          },
          "known": {
            "type": "boolean"
          },
          "lifecycleStatus": {
            "type": "string",
            "nullable": true
          },
          "authoritative": {
            "type": "boolean"
          },
          "relevantToManagedScope": {
            "type": "boolean"
          },
          "relevanceTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "seenBySources": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "seenInMempool": {
            "type": "boolean",
            "nullable": true
          },
          "blockHash": {
            "type": "string",
            "nullable": true
          },
          "blockHeight": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "firstSeenAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastObservedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "validationStatus": {
            "type": "string",
            "nullable": true
          },
          "payloadAvailable": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ValidateStasResponse": {
        "type": "object",
        "properties": {
          "askLater": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "isLegal": {
            "type": "boolean"
          },
          "isIssue": {
            "type": "boolean"
          },
          "isRedeem": {
            "type": "boolean"
          },
          "eventType": {
            "type": "string",
            "nullable": true
          },
          "spendingType": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "optionalDataContinuity": {
            "type": "boolean",
            "nullable": true
          },
          "tokenId": {
            "type": "string",
            "nullable": true
          },
          "roots": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "illegalRoots": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "validationStatus": {
            "type": "string",
            "nullable": true
          },
          "b2GResolved": {
            "type": "boolean"
          },
          "missingDependencies": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "validationRepairState": {
            "type": "string",
            "nullable": true
          },
          "validationRepairUpdatedAt": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}